requirements.txt
# Core AWS SDK
boto3>=1.34.0,<2.0.0
botocore>=1.34.0,<2.0.0

# Async support for better performance
aioboto3>=12.3.0,<13.0.0

# Retry and exponential backoff utilities
tenacity>=8.2.3,<9.0.0

# For parsing AWS ARNs and credential handling
python-dateutil>=2.8.2,<3.0.0

# Security and encryption
cryptography>=41.0.0,<43.0.0

# Database drivers (for future PostgreSQL support)
psycopg2-binary>=2.9.9,<3.0.0

# Type checking support
typing-extensions>=4.9.0,<5.0.0

# Microsoft Graph API support
msal>=1.25.0,<2.0.0
requests>=2.31.0,<3.0.0

# Google Workspace API support (future)
google-auth>=2.25.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.110.0,<3.0.0

